1. Изменения в классах api и БД. 2. Выделил проверку наличия города в БД и добавление нового города в БД в отдельные методы, а также написал событие, обновляющее combo box, когда добавляется новый город.

+++ Model/Database/City.cs
@@ -16,7 +16,7 @@ namespace WeatherForecastApp.Database
         public string Name { get; set; }
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
-        public int? Timezone { get; set; }
+        public ulong? Timezone { get; set; }
 
         public List<Forecast> Forecasts { get; set; }
     }
